chore: move form-theme component
