+++ src/components/form-theme/configured-form-theme.jsx
@@ -1,7 +1,7 @@
 import { reduxForm } from 'redux-form';
 
-import { FormThemes } from './form-theme.component';
-import { asyncValidateDatasetInvokePatch } from '../formsLib/asyncValidateDatasetInvokePatch';
+import FormThemes from './form-theme.component';
+import { asyncValidateDatasetInvokePatch } from '../../pages/dataset-registration-page/formsLib/asyncValidateDatasetInvokePatch';
 
 const config = {
   form: 'themes',
